+++ packages/g-plugin-canvaskit-renderer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @antv/g-plugin-canvaskit-renderer
 
+## 1.3.19
+
+### Patch Changes
+
+-   Updated dependencies [4fdee19f]
+-   Updated dependencies [4fdee19f]
+    -   @antv/g-plugin-image-loader@1.3.18
+    -   @antv/g-lite@1.2.18
+
 ## 1.3.18
 
 ### Patch Changes