+++ packages/g-plugin-mobile-interaction/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-mobile-interaction
 
+## 0.9.18
+
+### Patch Changes
+
+-   Updated dependencies [4fdee19f]
+-   Updated dependencies [4fdee19f]
+    -   @antv/g-lite@1.2.18
+
 ## 0.9.17
 
 ### Patch Changes