+++ packages/g-plugin-rough-svg-renderer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @antv/g-plugin-rough-svg-renderer
 
+## 1.9.23
+
+### Patch Changes
+
+-   Updated dependencies [4fdee19f]
+-   Updated dependencies [4fdee19f]
+    -   @antv/g-lite@1.2.18
+    -   @antv/g-svg@1.10.22
+
 ## 1.9.22
 
 ### Patch Changes